Bind foreground service to application
It was bound to an activity that could be destroyed at any point.
Android has been warning me about this in logcat for weeks.

Also took the opportunity to replace references with null. I noticed a
leak would occur if I unbind the service when record fragment pauses. I
do not think I have found the culprit but the leak disappeared after
setting those fields to null.